Change to int the type of expiration in class JwtUtil and change jwtsecret to 256 bits

+++ src/main/java/gcp/testSpringSecurity/studySpringSecurityJWT/util/JwtUtil.java
@@ -17,7 +17,7 @@ public class JwtUtil {
     private String secret;
 
     @Value("${jwt.expiration}")
-    private long expiration;
+    private int expiration;
 
     private Key getKey() {
         return Keys.hmacShaKeyFor(secret.getBytes());
